+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.11.4 / 5.66.4] - 2023-10-
 
+
+### Changed
+- improved lock all boxes behavioure [#3350](https://github.com/sandboxie-plus/Sandboxie/issues/3350)
+
 ### Fixed
 - fixed issue with unmounting protected images [#3347](https://github.com/sandboxie-plus/Sandboxie/issues/3347)
 - fixed Unable to turn off highlight in Sbie Messages search results [#3338](https://github.com/sandboxie-plus/Sandboxie/issues/3338)